Skip to next repository if 'git svn rebase' fails
The 'git svn rebase' command sometimes fails with the following error:

  > Error from SVN, (160006): Invalid filesystem revision number: No
  > such revision 123456

This happens rarely enough not to cause a real issue, but often enough
to make manual removal of the lock file annoying.

Instead of simply exiting and leaving behind a dangling lock file, shrug
it off and move on to the next repository.