Attempt to remove --whitespace=fix
This was added back when I created the script because some random commit
could not be applied to its package branch due to whitespace errors.

This is an attempt to get rid of it as it mangles leading whitespace
when tabs are used in PKGBUILDs or patches. We need to preserve
whitespace in source files in order for checksums to match.

Should a commit fail to apply in the future, it will need to be
troubleshooted again (while making sure whitespace is preserved).